scripts : fix next dev version calculation [no ci] (#1351)
This commit updates the release script to correctly calculate the next
development version after a release.

The motivation for this is that the script currently increments the new
version for the next development cycle but the version is already set to
the updated version so this should not increment again.